Add separate Header component test.
